+++ tests/testthat/test-runFragPipeAnalysis.R
@@ -63,6 +63,7 @@ test_that("runFragPipeAnalysis works", {
         interactiveGroupColumn = NULL,
         complexFDRThr = 0.1,
         maxNbrComplexesToPlot = Inf,
+        maxComplexSimilarity = 1,
         seed = 123,
         includeFeatureCollections = "complexes",
         minSizeToKeepSet = 2,
@@ -568,6 +569,15 @@ test_that("runFragPipeAnalysis works", {
     expect_error(do.call(runFragPipeAnalysis, args),
                  "'maxNbrComplexesToPlot' must be within [0,Inf] (inclusive)",
                  fixed = TRUE)
+
+    ## maxComplexSimilarity
+    args <- args0
+    args$maxComplexSimilarity <- "1"
+    expect_error(do.call(runFragPipeAnalysis, args),
+                 "'maxComplexSimilarity' must be of class 'numeric'")
+    args$maxComplexSimilarity <- c(0.1, 0.2)
+    expect_error(do.call(runFragPipeAnalysis, args),
+                 "'maxComplexSimilarity' must have length 1")
 
     ## seed
     args <- args0