+++ tests/testthat/test-runMaxQuantAnalysis.R
@@ -67,6 +67,7 @@ test_that("runMaxQuantAnalysis works", {
         interactiveGroupColumn = NULL,
         complexFDRThr = 0.1,
         maxNbrComplexesToPlot = Inf,
+        maxComplexSimilarity = 1,
         seed = 42,
         includeFeatureCollections = "complexes",
         minSizeToKeepSet = 2,
@@ -593,6 +594,15 @@ test_that("runMaxQuantAnalysis works", {
     expect_error(do.call(runMaxQuantAnalysis, args),
                  "'maxNbrComplexesToPlot' must be within [0,Inf] (inclusive)",
                  fixed = TRUE)
+
+    ## maxComplexSimilarity
+    args <- args0
+    args$maxComplexSimilarity <- "1"
+    expect_error(do.call(runMaxQuantAnalysis, args),
+                 "'maxComplexSimilarity' must be of class 'numeric'")
+    args$maxComplexSimilarity <- c(0.1, 0.2)
+    expect_error(do.call(runMaxQuantAnalysis, args),
+                 "'maxComplexSimilarity' must have length 1")
 
     ## seed
     args <- args0