+++ tests/testthat/test-runPDTMTAnalysis.R
@@ -77,6 +77,7 @@ test_that("runPDTMTAnalysis works", {
         interactiveGroupColumn = NULL,
         complexFDRThr = 0.1,
         maxNbrComplexesToPlot = Inf,
+        maxComplexSimilarity = 1,
         seed = 42,
         includeFeatureCollections = "complexes",
         minSizeToKeepSet = 2,
@@ -689,6 +690,15 @@ test_that("runPDTMTAnalysis works", {
     expect_error(do.call(runPDTMTAnalysis, args),
                  "'maxNbrComplexesToPlot' must be within [0,Inf] (inclusive)",
                  fixed = TRUE)
+
+    ## maxComplexSimilarity
+    args <- args0
+    args$maxComplexSimilarity <- "1"
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'maxComplexSimilarity' must be of class 'numeric'")
+    args$maxComplexSimilarity <- c(0.1, 0.2)
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'maxComplexSimilarity' must have length 1")
 
     ## seed
     args <- args0